Add ability to specify a folder of configuration files
For some databases it may be easier to manage everything in separate files (schema, table objects, etc.). Allow the developer to specify a folder that contains all of the files. Any YAML files in that folder will be read in, concatenated into a single YAML file, and used to create the database object.